+++ module/actor/entity.js
@@ -978,7 +978,7 @@ export class ActorPF extends Actor {
     }
 
     const allowed = Hooks.call("actorRoll", this, "skill", skillId, options);
-    if ( allowed === false ) return;
+    if (allowed === false) return;
 
     let skl, sklName;
     const data = this.data.data,
@@ -1026,7 +1026,6 @@ export class ActorPF extends Actor {
     // Wound Threshold penalty
     const wT = this.getWoundThresholdData();
     if (wT.multiplier > 0 && wT.penalty > 0) {
-      mods.push(-wT.penalty);
       notes.push(game.i18n.localize(CONFIG.PF1.woundThresholdConditions[wT.level]));
     }
 
@@ -1070,7 +1069,7 @@ export class ActorPF extends Actor {
     }
 
     const allowed = Hooks.call("actorRoll", this, "bab", null, options);
-    if ( allowed === false ) return;
+    if (allowed === false) return;
 
     return DicePF.d20Roll({
       event: options.event,
@@ -1092,7 +1091,7 @@ export class ActorPF extends Actor {
     }
 
     const allowed = Hooks.call("actorRoll", this, "cmb", null, options);
-    if ( allowed === false ) return;
+    if (allowed === false) return;
 
     // Add contextual notes
     let notes = [];
@@ -1133,7 +1132,7 @@ export class ActorPF extends Actor {
     rollData.cl = spellbook.cl.total;
 
     const allowed = Hooks.call("actorRoll", this, "cl", spellbookKey, options);
-    if ( allowed === false ) return;
+    if (allowed === false) return;
 
     // Add contextual caster level string
     const notes = this.getContextNotesParsed(`spell.cl.${spellbookKey}`);
@@ -1165,7 +1164,7 @@ export class ActorPF extends Actor {
     rollData.concentrationBonus = spellbook.concentration;
 
     const allowed = Hooks.call("actorRoll", this, "concentration", spellbookKey, options);
-    if ( allowed === false ) return;
+    if (allowed === false) return;
 
     // Add contextual concentration string
     const notes = this.getContextNotesParsed(`spell.concentration.${spellbookKey}`);
@@ -1320,7 +1319,7 @@ export class ActorPF extends Actor {
     }
 
     const allowed = Hooks.call("actorRoll", this, "save", savingThrowId, options);
-    if ( allowed === false ) return;
+    if (allowed === false) return;
 
     // Add contextual notes
     let notes = [];
@@ -1382,7 +1381,7 @@ export class ActorPF extends Actor {
     }
 
     const allowed = Hooks.call("actorRoll", this, "ability", abilityId, options);
-    if ( allowed === false ) return;
+    if (allowed === false) return;
 
     // Add contextual notes
     let notes = [];
@@ -2225,11 +2224,6 @@ export class ActorPF extends Actor {
         }
         break;
     }
-
-    // Wound Threshold penalty shorthand
-    const wT = this.getWoundThresholdData(result);
-    setProperty(result, "attributes.woundThresholds.penaltyBase", wT.level * wT.multiplier); // To aid relevant formulas
-    setProperty(result, "attributes.woundThresholds.penalty", wT.penalty);
 
     return result;
   }